+++ wwpdb/apps/workmanager/task_access/SequenceMerger.py
@@ -37,7 +37,7 @@ class SequenceMerger(BaseClass):
     def run(self):
         """
         """
-        numProc = multiprocessing.cpu_count() / 2
+        numProc = int(multiprocessing.cpu_count() / 2)
         mpu = MultiProcUtil(verbose = True)
         mpu.set(workerObj = self, workerMethod = "runMulti")
         mpu.setWorkingDir(self._sessionPath)
